+++ python/marvin/utils/plot/__init__.py
@@ -1,1 +1,2 @@
 from marvin.utils.plot.colorbar import *
+from .utils import *